+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Yii2 Admin Change Log
 1.0.4 Under Development
 -----------------------
 
+- Bug: #102: Unique validation of the permission and role (mdmunir).
 - Enh: Clear cache when menu change (mdmunir).
 - Enh: Ensure get latest state of `user` component (mdmunir).